requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
Pillow>=10.4.0
numpy>=1.26.4
openai>=1.30.1
tensorflow>=2.17.0
pydantic>=2.7.1
aiofiles>=23.2.1